# app/api/vulnerability_api.py
from flask import Blueprint, jsonify, request

from ..error_handler import ValidationError
from ..service.vulnerability_service import VulnerabilityService

vulnerability_bp = Blueprint("vulnerability", __name__)
service = VulnerabilityService()


@vulnerability_bp.route("/snapshots", methods=["POST"])
def create_snapshot():
    data = request.get_json(force=True)
    idempotency_key = request.headers.get("Idempotency-Key")
    result, status_code = service.create_snapshot(data, idempotency_key=idempotency_key)
    return jsonify(result), status_code


@vulnerability_bp.route("/snapshots/async", methods=["POST"])
def create_snapshot_async():
    data = request.get_json(force=True)
    idempotency_key = request.headers.get("Idempotency-Key")
    job_id = service.submit_snapshot(data, idempotency_key=idempotency_key)
    return jsonify({"job_id": job_id}), 202


@vulnerability_bp.route("/snapshots/<job_id>/status", methods=["GET"])
def get_snapshot_status(job_id):
    status = service.get_snapshot_status(job_id)
    return jsonify({"status": status}), 200


@vulnerability_bp.route("/snapshots/<job_id>/result", methods=["GET"])
def get_snapshot_result(job_id):
    result = service.get_snapshot_result(job_id)
    return jsonify(result), 200


@vulnerability_bp.route(
    "/products/<product_name>/versions/<product_version>/snapshots", methods=["GET"]
)
def list_snapshots(product_name, product_version):
    try:
        limit = min(int(request.args.get("limit", 10)), 100)
        offset = int(request.args.get("offset", 0))
    except ValueError:
        raise ValidationError(
            details=[{"field": "query", "message": "limit and offset must be integers"}]
        )

    if limit < 0 or offset < 0:
        raise ValidationError(
            details=[{"field": "query", "message": "limit and offset must be non-negative"}]
        )

    result = service.list_snapshots(product_name, product_version, limit, offset)
    return jsonify(result), 200


@vulnerability_bp.route("/snapshots/<snapshot_id>", methods=["GET"])
def get_snapshot(snapshot_id):
    result = service.get_snapshot(snapshot_id)
    return jsonify(result), 200


@vulnerability_bp.route("/snapshots/<snapshot_id>/changes", methods=["GET"])
def get_snapshot_changes(snapshot_id):
    try:
        limit = min(int(request.args.get("limit", 10)), 100)
        offset = int(request.args.get("offset", 0))
    except ValueError:
        raise ValidationError(
            details=[{"field": "query", "message": "limit and offset must be integers"}]
        )

    if limit < 0 or offset < 0:
        raise ValidationError(
            details=[{"field": "query", "message": "limit and offset must be non-negative"}]
        )

    filters = {k: v for k, v in request.args.items() if k not in ("limit", "offset")}
    result = service.get_snapshot_changes(snapshot_id, filters, limit, offset)
    return jsonify(result), 200